Fixed errors in primt_memory, debug, and conversion

+++ include/common/conversion.h
@@ -50,7 +50,7 @@ uint8_t my_itoa(int32_t data, uint8_t* ptr, uint32_t base);
  *
  * @return integer converted from ASCII string
  */
-uint8_t my_atoi(uint8_t* ptr, uint8_t digits, uint32_t base);
+int32_t my_atoi(uint8_t* ptr, uint8_t digits, uint32_t base);
 
 /**
  * @brief function to convert data from big endian to little endian
